// app/types/movieType.ts
export interface Person {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
}

export interface CastMember extends Person {
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface CrewMember extends Person {
  credit_id: string;
  department: string;
  job: string;
}

export interface Genre {
  id: number;
  name: string;
}

export interface ProductionCompany {
  id: number;
  logo_path: string;
  name: string;
  origin_country: string;
}

export interface ProductionCountry {
  iso_3166_1: string;
  name: string;
}

export interface SpokenLanguage {
  english_name: string;
  iso_639_1: string;
  name: string;
}

export interface VideoResult {
  iso_639_1: string;
  iso_3166_1: string;
  name: string;
  key: string;
  site: string;
  size: number;
  type: string;
  official?: boolean;
  published_at: string;
  id: string;
}

export interface AuthorDetails {
  name: string;
  username: string;
  avatar_path: string;
  rating: string;
}